Fix Error: ./src/app/sessions You cannot have two parallel pages that resolve to the same path.
